Datasets/requirements.txt
numpy>=1.21
pandas>=1.3
librosa>=0.9.2
pydub>=0.25.1
scipy>=1.7
soundfile>=0.10.3
torchaudio>=0.10
torch>=1.10
tqdm>=4.62
PyYAML>=6.0
audiomentations>=0.30